:recycle: reorganizar modelos
